Use correct OS division of code for compile reasons
